Move Internal.Console to shared CoreLib & port to Unix
