+++ src/libraries/Common/src/Interop/Unix/System.Native/Interop.Log.cs
