+++ src/libraries/Native/Unix/System.Native/CMakeLists.txt
@@ -28,7 +28,9 @@ if (CLR_CMAKE_TARGET_IOS OR CLR_CMAKE_TARGET_TVOS)
         pal_log.m
         pal_searchpath.m)
 else ()
-    set(NATIVE_SOURCES ${NATIVE_SOURCES} pal_console.c)
+    set(NATIVE_SOURCES ${NATIVE_SOURCES}
+        pal_console.c
+        pal_log.c)
 endif ()
 
 if (CLR_CMAKE_TARGET_LINUX AND NOT CLR_CMAKE_TARGET_BROWSER)
