+++ src/libraries/System.Console/src/System.Console.csproj
@@ -21,8 +21,8 @@
   <!-- iOS/tvOS -->
   <ItemGroup Condition="'$(TargetsiOS)' == 'true' or '$(TargetstvOS)' == 'true'">
     <Compile Include="System\ConsolePal.iOS.cs" />
-    <Compile Include="$(CommonPath)Interop\OSX\System.Native\Interop.Log.cs"
-             Link="Common\Interop\OSX\Interop.Log.cs" />
+    <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.Log.cs"
+             Link="Common\Interop\Unix\Interop.Log.cs" />
     <Compile Include="$(CommonPath)Interop\Unix\Interop.Libraries.cs"
              Link="Common\Interop\Unix\Interop.Libraries.cs" />
   </ItemGroup>
